+++ src/Handlers/SwishPaymentHandler.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Buckaroo\Shopware6\Handlers;
+
+use Buckaroo\Shopware6\PaymentMethods\Swish;
+
+class SwishPaymentHandler extends PaymentHandlerSimple
+{
+    public string $paymentClass = Swish::class;
+}