+++ src/Helpers/GatewayHelper.php
@@ -34,6 +34,9 @@ use Buckaroo\Shopware6\PaymentMethods\Creditcards;
 use Buckaroo\Shopware6\PaymentMethods\PayPerEmail;
 use Buckaroo\Shopware6\PaymentMethods\Knaken;
 use Buckaroo\Shopware6\PaymentMethods\SepaDirectDebit;
+use Buckaroo\Shopware6\PaymentMethods\Swish;
+use Buckaroo\Shopware6\PaymentMethods\Bizum;
+use Buckaroo\Shopware6\PaymentMethods\Twint;
 
 class GatewayHelper
 {
@@ -67,6 +70,9 @@ class GatewayHelper
         IdealQr::class,
         MBWay::class,
         Multibanco::class,
-        Knaken::class
+        Knaken::class,
+        Swish::class,
+        Bizum::class,
+        Twint::class
     ];
 }