+++ src/Resources/app/administration/src/components/buckaroo-payment-list/index.js
@@ -137,6 +137,18 @@ Component.register("buckaroo-payment-list", {
                 {
                     code: "WeChatPay",
                     logo: "wechatpay.svg"
+                },
+                {
+                    code: "swish",
+                    logo: "swish.svg"
+                },
+                {
+                    code: "bizum",
+                    logo: "bizum.svg"
+                },
+                {
+                    code: "twint",
+                    logo: "twint.svg"
                 }
             ]
         };